backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pymupdf==1.23.8
python-multipart==0.0.6
requests==2.31.0